refactor(moyo): bake aperiodic-c origin-shift correction into integral_normalizer_2_1
`LayerGroup::new` previously had a `// TODO: refactor` block that
overrode the c-component of the origin shift after the normalizer
returned, because `match_origin_shift` -> `solve_mod1` reduces mod 1
on all three components even though the layer cell's c-axis is
aperiodic. Move that override inside `integral_normalizer_2_1` so
every returned `UnimodularTransformation` already carries an exact,
non-mod-1 `s_z`. The `layer_exact_s_z` helper moves from
`identify/layer_group.rs` to `identify/normalizer.rs` as a
file-private fn alongside its only caller.

`LayerGroup::new` collapses to: project rotations, match arithmetic
class, iterate candidate Halls, and return the first conjugator from
`integral_normalizer_2_1` directly. The static
`LAYER_CORRECTION_MATRICES` rationale and Stage 2 / Stage 3 split in
`docs/layer_architecture.md` are reworded to reflect the merged
two-stage flow.

Pure refactor: bit-for-bit equivalent for the first conjugator
(which is what the caller consumes). All 178 cargo tests pass with
zero diffs in identified `(hall_number, transformation.linear,
transformation.origin_shift)`, including the three round-trip pins
`test_round_trip_all_layer_hall_numbers`,
`test_round_trip_default_hall_numbers_via_standard_setting`, and
`test_normalizer_aligns_non_canonical_basis`.
